+++ packages/nano-ui-markdown/examples/Main.hs
@@ -1,7 +1,7 @@
 module Main (main) where
 
 import ChatDemo (chatDemoUi)
-import NanoUI (Key (KeyEscape), Size (..), WindowSettings (..), defaultWindowSettings, inputKeys, inputKeysElem)
+import NanoUI (Key (KeyEscape), Pressable (..), Size (..), WindowSettings (..), defaultWindowSettings)
 import NanoUI.Backend.Sdl
   ( SdlOptions (..)
   , defaultSdlOptions
@@ -13,6 +13,6 @@ main =
   runSdlApp
     defaultSdlOptions
       { sdlWindowSettings = defaultWindowSettings {wsTitle = "nano-ui-markdown example", wsSize = Size 900 760}
-      , sdlAppShouldQuit = \inp -> inputKeysElem KeyEscape (inputKeys inp)
+      , sdlAppShouldQuit = pressedOnceIn KeyEscape
       }
     chatDemoUi